Fix for exit intent modal blocking clicks
The exit intent modal was blocking clicks on the main webpage, even when it was dismissed.

This is because the HTML for the modal was injected on top of the content and when dismissed was just hidden, not completely removed, meaning that clicks elsewhere on the webpage were blocked by the invisible element.

I have fixed this by adding CSS to the overlay to pass click events through it when it is in the .fade-out state, and recapture those click events when it is in the .fade-in state.