Update config.json
Worked with Alondra Guzman from Quantum Metric to update the integration in their docs.

The original integration used a getDecisionObject call to get some values to pass into the QM integration.

They were not aware of the availablity of standard variables in the integration, there was no need to grab the decision object.

Further, the decision object is different when returning an experiment than when a personalization. There was no check for this and it was causing an error in certain scenarios.